refactor(Session): fix Action Manager things in session
I may have to correct something about Action Manager and its classes: ActionDescriptor, ActionList, ActionReference. Action Manager is NOT something to describe your action palette. Action Manager actually enables low-level access into photoshop, and all things in this repo can actually be rewritten in Action Manager. You can consider an ActionDescriptor as dict, ActionList as array, ActionReference as filters for searching. So we should NOT set an INSTANCE of ActionDescriptor as an attribute of a Session and instead we should set the ActionDescriptor CLASS (the same for ActionList, ActionReference).

Signed-off-by: TsXor <zhang050525@qq.com>