fix(action_manager/utils): 1. Requires initialization of app when needed instead of do it on import to pass import test.  2. Warns user when typeID collision happens. (And id2str will try to avoid giving user string that will cause collision.)
typeID collision happens when a stringID and a charID share the same name. For example, charID 'From' -> stringID 'from' and charID 'from' -> stringID 'originalAddressAttr', so 'from' is a string that will cause collision. After scanning the cpp header "PITerminology.h" and "PIStringTerminology.h", I found 3 such strings and hardcoded them into utils.py. str2id() will not automatically correct user when user gives a collision string (because no one knows what he want it to be a stringID or an charID!), but will regard that string as stringID and throw a warning. Users can follow that warning's guide to fix error and suppress that warning.

Signed-off-by: TsXor <zhang050525@qq.com>