refactor(Photoshop): Restructure the sequence of events attempting to initialize a valid Dispatch object
The goal of restructuring is to hopefully reduce the recent issues arising from broken or mismatched registry data that results in the API failing to establish a valid Dispatch object. These recent occurences appear to coincide with Photoshop's rollover to v25 and the occurence of mixed registry data between Photoshop and Photoshop beta. This commit also attempts to refactor for readability and performance. Additionally: I deprecated the use of `GetActiveObject`, at the time I introduced it I didn't realize this spawner *only* seems to work for the initial Application object, so spawning any other object (ActionDescriptor, SolidCOlor, etc) just pads the execution time with zero benefit. We should definitely cut it loose.

Signed-off-by: MrTeferi <freethoughtleft@gmail.com>